requirements.txt
black==21.10b0
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ipython==7.29.0
isort==5.10.0
Pillow==8.4.0
PyJWT==2.3.0